backend/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
chromadb==1.3.4
google-genai==1.49.0
python-dotenv==0.21.0
tenacity==9.1.2
httpx==0.28.1